backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.3.0
torch>=2.0.0,<2.6.0
torchvision>=0.15.0,<0.21.0
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
werkzeug==3.0.1
supabase==2.3.0
python-dotenv==1.0.0
openai==1.12.0
PyPDF2==3.0.1

